changed for working version

+++ database/migrations/2018_12_03_154613_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->increments('id');
             $table->string('title');
             $table->text('content');
-            $table->decimal('price', 2);
+            $table->float('price');
             $table->timestamps();
         });
     }
